MemoryManager/Memory Allocaters Update
ADD:
    - StackMemoryAllocater: a stack allocater which extends the memory allocater functionality (uses 4-bytes more memory for every allocation)
    - StackMemoryAllocaterE: an efficient stack allocater which safes memory by not keeping track of the allocations being made; to free the top of stack, users have to provide the pointer to the last allocated memory

UPDATE:
    - MemoryAllocater: Re-Implemented the class, just a base class allocaters will need to extend
    - MemoryManager: now only offers aligned memory
       - MallocAllocaterAligned(): returns a aligned allocated memory allocater
    - MemoryPool: minor changes
       - Malloc(): changed param from uint64 to uint128

+++ Source/Runtime/Memory/Core/MemoryPool.h
@@ -70,7 +70,7 @@ public:
 	*
 	* @return char* pointer pointing to the memory location
 	*/
-	char* Malloc(uint64 inSizeInBytes)
+	char* Malloc(uint128 inSizeInBytes)
 	{
 		// Check if we can allocate enough memory
 #if _DEBUG | _EDITOR
